refactor: extract pipeline helpers, label specs, and options from train.py (#74)
